Split out SMP tests into their own folder.
Will make it easier to add other tests, like sharing the GIC across
cores.
