+++ examples/mps3-an536-smp/rust-toolchain.toml
@@ -1,0 +1,6 @@
+[toolchain]
+channel = "nightly-2026-01-26"
+targets = [
+	"armv8r-none-eabihf",
+]
+components = ["rust-src", "clippy", "rustfmt"]
