--- examples/mps3-an536/reference/smp_test-armv8r-none-eabihf_smp2.out
@@ -1,2 +1,0 @@
-CAS test passed
-CS Mutex test passed
