--- examples/mps3-an536/reference/smp_test-thumbv8r-none-eabihf.out
@@ -1,1 +1,0 @@
-CPU 1 is missing?!
